Merged Pin and CompressiblePin
Therefore removed `BARRIER_PIN`. Barrier pins were the only pin type which
didn't have a button so they got removed in order to make the design less ugly.
At this point in time, there is no `Pin` that is not also a `ButtonPin`. These
will eventually be merged in the future.